Updated UI and features
Added the ability to adapt to different resolutions on the fly. Also, implemented a context menu when right clicking the hidden icon. The context menu include the feature of turning on the locky keys while giving the user the options such as running the application on startup and exiting the application. Also name change to lock switch.